Renaming files to be more explicit with the includes (without relying on directory names)
