+++ operators/dds/video/dds_video_subscriber/python/dds_video_subscriber.cpp
@@ -21,8 +21,8 @@
 #include <memory>
 #include <string>
 
-#include "../subscriber.hpp"
-#include "./subscriber_pydoc.hpp"
+#include "../dds_video_subscriber.hpp"
+#include "./dds_video_subscriber_pydoc.hpp"
 
 #include "../../../../operator_util.hpp"
 #include "holoscan/core/fragment.hpp"
